+++ snf4j-core/src/test/java/org/snf4j/core/DTLSSessionTest.java
@@ -107,6 +107,7 @@ public class DTLSSessionTest extends DTLSTest {
 		c.startClient();
 		c.waitForSessionReady(TIMEOUT);
 		s.waitForSessionReady(TIMEOUT);
+		waitFor(50);
 		c.getSession().close();
 		c.waitForSessionEnding(TIMEOUT);
 		s.waitForSessionEnding(TIMEOUT);
@@ -139,6 +140,7 @@ public class DTLSSessionTest extends DTLSTest {
 		c.startClient();
 		c.waitForSessionReady(TIMEOUT);
 		s.waitForSessionReady(TIMEOUT);
+		waitFor(50);
 		c.stop(TIMEOUT);
 		c.waitForSessionEnding(TIMEOUT);
 		s.waitForSessionEnding(TIMEOUT);
@@ -150,6 +152,7 @@ public class DTLSSessionTest extends DTLSTest {
 		c.startClient();
 		c.waitForSessionReady(TIMEOUT);
 		s.waitForSessionReady(TIMEOUT);
+		waitFor(50);
 	}
 	
 	@Test